\docType{data}
\name{aeroportos}
\alias{aeroportos}
\title{Aeroportos}
\format{Um data.frame com 1.458 linhas e 8 colunas
\describe{
\item{codigo_aeroporto}{Código FAA do aeroporto}
\item{nome}{Nome do aeroporto}
\item{latitude}{Coordenada de localização do aeroporto}
\item{longitude}{Coordenada de localização do aeroporto}
\item{altura}{Altura medida em pés}
\item{fuso_horario}{Diferença de fuso horário do Meridiano de Greenwich (GMT)}
\item{horario_verao}{Horário legal de acordo com as políticas de economia de energia. A = Horário de verão (Daylight Saving Time) padrão dos EUA: começa o segundo sábado de março E; termina o primeiro sábado de novembro; U = Desconhecido; N = Não possui horário de verão}
\item{fuso_horario_iana}{Fuso horário IANA, determinado de acordo com o serviço web GeoNames}
}}
\usage{aeroportos}
\description{Metadados úteis sobre aeroportos}
\keyword{datasets}
