\docType{data}
\name{avioes}
\alias{avioes}
\title{Dados de aviões}
\format{Um data.frame com 3.322 linhas e 9 colunas
\describe{
\item{codigo_cauda}{Código da cauda do avião (empenagem)}
\item{ano}{Ano de fabricação}
\item{tipo}{Tipo de avião}
\item{fabricante}{Fabricante do avião}
\item{modelo}{Modelo de avião}
\item{motores}{Quantidade de motores}
\item{assentos}{quantidade de assentos}
\item{velocidade}{Velocidade média de cruzeiro em milhas por hora}
\item{tipo_motor}{Tipo de motor do avião}
}}
\usage{avioes}
\description{Metadatos de aviões e seus códigos de cauda no registro de aviação da FAA. American Airways (AA) e Envoy Air (MQ) reportam número de frota no lugar do número de calda, por isso não é possível relacioná-los}
\keyword{datasets}
