\docType{data}
\name{comuns}
\alias{comuns}
\title{Modelos comuns de carros}
\format{Um data.frame com 347 linhas e 4 colunas
\describe{
\item{marca}{Marca}
\item{modelo}{Modelo}
\item{total_modelos}{Número total de modelos}
\item{total_modelos_ano}{Número total de modelos-ano}
}}
\usage{comuns}
\description{Esses modelos têm pelo menos 10 anos de dados. Esse conjunto de dados é adequado para uma join para restringir o número de observaçôes a um nível mais gerenciável.}
\seealso{\code{\link[fueleconomy]{common}}}
\keyword{datasets}
