% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringr.R
\docType{data}
\name{dados-stringr}
\alias{dados-stringr}
\alias{frases}
\alias{frutas}
\alias{palavras}
\title{Vetores de caracteres do pacote stringr}
\format{
Um vetor

An object of class \code{character} of length 37.

An object of class \code{character} of length 116.

An object of class \code{character} of length 267.
}
\usage{
frases

frutas

palavras
}
\description{
Amostra de vetores de caracteres para praticar a manipulação de strings
}
\details{
frases - é uma lista de frases escritas em português

frutas - é uma seleção e adaptação ao português dos dados de prática contidos em stringr::fruit.

palavras - é uma lista de palavras frequentes usadas em português brasileiro
}
\examples{

length(frases)
frases[1:5]

length(frutas)
frutas[1:5]

#' length(palavras)
palavras[1:5]
}
\keyword{datasets}
