\docType{data}
\name{dados_gapminder}
\alias{dados_gapminder}
\title{Dados de Gapminder}
\format{Um data.frame de 1.704 linhas e 6 colunas
\describe{
\item{pais}{Países incluídos (fator com 142 níveis)}
\item{continente}{Continentes (fator com 5 níveis)}
\item{ano}{De 1952 a 2007, dados a cada 5 anos}
\item{expectativa_de_vida}{Expectativa de vida ao nascer, em anos}
\item{populacao}{População}
\item{pib_per_capita}{PIB per capita (em dólares americanos, ajustados pela inflação)}
}}
\usage{dados_gapminder}
\description{Conjunto de dados de Gapminder sobre expectativa de vida, PIB per capita e população por país}
\source{http://www.gapminder.org/data/}
\keyword{datasets}
