\docType{data}
\name{dados_iris}
\alias{dados_iris}
\title{Dados de espécies de flor de Íris por Edgar Anderson}
\format{Um data.frame com 150 linhas e 5 colunas
\describe{
\item{Comprimento.Sepala}{Comprimento da Sépala}
\item{Largura.Sepala}{Largura da Sépala}
\item{Comprimento.Petala}{Comprimento da Pétala}
\item{Largura.Petala}{Largura da Pétala}
\item{Especies}{A espécie correspondente da flor de Íris (Iris setosa, I. versicolor, I. virginica)}
}}
\usage{dados_iris}
\description{O famoso conjunto de dados sobre espécies de flor de Íris (Edgar Anderson e Ronald Fisher) fornece as medidas em centímetros das variáveis de comprimento e largura da sépala e pétala, para 50 flores de cada uma das 3 espécies de íris. As espécies são Iris setosa, I. versicolor e I. virginica}
\keyword{datasets}
