% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_pessoas.R
\name{gerar_pessoas}
\alias{gerar_pessoas}
\title{Gera dados aleatórios sobre pessoas físicas do 4devs.}
\usage{
gerar_pessoas(n = 5)
}
\arguments{
\item{n}{Número de observações. Máximo 30 pessoas.}
}
\value{
data.frame
}
\description{
Gera dados aleatórios sobre pessoas físicas do 4devs.
}
\examples{
df <- gerar_pessoas(n = 5)
}
