\docType{data}
\name{mtcarros}
\alias{mtcarros}
\title{Testes de estrada para automóveis}
\format{Um data.frame com 32 linhas e 12 colunas
\describe{
\item{milhas_por_galao}{Milhas por galão (US)}
\item{cilindros}{Número de cilindros}
\item{cilindrada}{Volume de deslocamento do motor em polegadas cúbicas}
\item{cavalos_forca}{Cavalos força bruto}
\item{eixo}{Relação de eixo traseiro}
\item{peso}{Peso (1000 libras)}
\item{velocidade}{Tempo em percorrer 1/4 de milha}
\item{forma}{Forma do motor (V ou em linha)}
\item{transmissao}{Tipo de transmissão (0 = automático, 1 = manual)}
\item{marchas}{Número de marchas à frente}
\item{carburadores}{Número de carburadores}
}}
\usage{mtcarros}
\description{Os dados foram extraídos da revista Motor Trend US, de 1974, e incluem o consumo de combustível e 10 aspectos do design e desempenho de automóveis para 32 automóveis (modelos 1973-74)}
\keyword{datasets}
