\docType{data}
\name{nucleo_familiar}
\alias{nucleo_familiar}
\title{Dados de núcleo familiar}
\format{Um tibble com 5 linhas e 5 colunas
\describe{
\item{familia}{Identificador da família}
\item{datanascimento_crianca1}{Data de nascimento da primeira criança}
\item{datanascimento_crianca2}{Data de nascimento da segunda criança}
\item{nome_crianca1}{Nome da primeira criança}
\item{nome_crianca2}{Nome da segunda criança}
}}
\description{Contém observações de 5 famílias, com até 2 crianças cada. Esses dados são baseados no exemplo disponível em: vignette('datatable-reshape', package = 'data.table')}
\keyword{datasets}
