% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_ofertas_ml.R
\name{obter_ofertas_ml}
\alias{obter_ofertas_ml}
\title{Obter ofertas mais recentes do Mercado Livre}
\usage{
obter_ofertas_ml(paginas = 1)
}
\arguments{
\item{paginas}{Número de páginas}
}
\value{
Tibble
}
\description{
Obter ofertas mais recentes do Mercado Livre
}
\examples{
df <- obter_ofertas_ml(paginas = 1)
}
