\docType{data}
\name{pinguins}
\alias{pinguins}
\title{Medidas de pinguins adultos perto da Estação Palmer, Antártida (Palmer Station)}
\format{Um tibble com 344 linhas e 8 colunas
\describe{
\item{especie}{Um fator com as espécies de pinguim (Pinguim-de-adélia, Pinguim-de-barbicha e Pinguim-gentoo)}
\item{ilha}{Um fator com cada ilha do Arquipélago Palmer, na Antártida (Biscoe, Dream, Togersen)}
\item{comprimento_bico}{Um número inteiro que indica o comprimento do bico (em milímetros)}
\item{profundidade_bico}{Um número inteiro que indica a profundidade do bico (em milímetros)}
\item{comprimento_nadadeira}{Um número inteiro que indica o comprimento da nadadeira (em milímetros)}
\item{massa_corporal}{Um número inteiro que indica a massa corporal (em gramas)}
\item{sexo}{Um fator que indica o sexo do(a) pinguim (macho, fêmea)}
\item{ano}{Um número inteiro que indica o ano de estudo (2007, 2008 ou 2009)}
}}
\description{Inclui espécies de pinguins e ilhas do Arquipélago Palmer, medidas de cada espécie (comprimento da nadadeira, massa corporal e dimensões do bico), sexo de cada pinguim e ano de documentação. Esse é um subconjunto de dados do `penguins_raw`.}
\keyword{datasets}
