\docType{data}
\name{pixar_generos}
\alias{pixar_generos}
\title{Gêneros dos filmes da Pixar}
\format{Um tibble com 128 linhas e 2 colunas
\describe{
\item{filme}{Título do filme}
\item{genero}{Gênero do filme}
}}
\usage{pixar_generos}
\description{Filmes e seus gêneros}
\seealso{\code{\link[pixarfilms]{genres}}}
\keyword{datasets}
