\docType{data}
\name{pixar_indicacoes_oscar}
\alias{pixar_indicacoes_oscar}
\title{Filmes da Pixar com indicações ao Oscar}
\format{Um tibble com 72 linhas e 3 colunas
\describe{
\item{filme}{Título do filme}
\item{tipo_premio_indicado}{Tipo de prêmio}
\item{resultado}{Resultado da indicação}
}}
\usage{pixar_indicacoes_oscar}
\description{Filmes da Pixar que foram indicados ao Oscar, a categoria em que concorreram e o resultado da indicação}
\seealso{\code{\link[pixarfilms]{academy}}}
\keyword{datasets}
