\docType{data}
\name{premios_gerentes}
\alias{premios_gerentes}
\title{Tabela de prêmios dos gerentes}
\format{Um data.frame com 179 linhas e 6 colunas
\describe{
\item{id_gerente}{ID do gerente}
\item{nome_premio}{Nome do prêmio recebido pelo gerente}
\item{ano}{Ano}
\item{id_liga}{ID liga (fator)}
\item{empate}{Tem valor "S" (sim) para os prêmios que correspondem a um empate e "N" (não) caso contrário.}
\item{notas}{Notas sobre o prêmio}
}}
\description{Informações sobre prêmios para gerentes}
\keyword{datasets}
