\docType{data}
\name{rebatedores}
\alias{rebatedores}
\title{Tabela de estatíticas de Beisebol}
\format{Um data frame com 105.861 linhas e 22 colunas
\describe{
\item{id_jogador}{ID do jogador}
\item{id_ano}{Ano}
\item{ordem_equipes}{Ordem na qual o jogador se mudou entre equipes na mesma temporada}
\item{id_equipe}{ID da equipe (fator)}
\item{id_liga}{ID da liga (fator com os níveis AA, AL, FL, NL, PL, UA)}
\item{jogos}{Número de jogos em que o jogador participou oficialmente}
\item{vez_bastao}{Número de vezes ao bastão válidas}
\item{corridas}{Corridas anotadas}
\item{rebatidas_simples}{Número de vezes em que o jogador chegou à primeira base com uma rebatida válida}
\item{rebatidas_duplas}{Número de vezes em que o jogador chegou à segunda base com uma rebatida válida}
\item{rebatidas_triplas}{Número de vezes em que o jogador chegou à terceira base com uma rebatida válida}
\item{homerun}{Número de homeruns}
\item{corridas_impulsionadas}{Corridas impulsionadas como resultado da vez ao bastão do rebatedor}
\item{bases_roubadas}{Bases roubadas}
\item{capturado_roubando}{Eliminações em tentativa de roubo de base}
\item{base_bolas}{Base por bolas}
\item{strikeouts}{Eliminações por strikes}
\item{base_intencional}{Base por bolas intencional}
\item{HBP}{Vezes atingidas pelo arremessador (Hit by pitch - HBP)}
\item{toque_sacrificio}{Toque de sacrifício}
\item{fly_sacrificio}{Flys (rebatidas altas no campo externo) de sacrifício (impulsiona uma corrida)}
\item{eliminacao_dupla_induzida}{Rebatidas rasteiras induzindo a uma eliminação dupla}
}}
\description{Estatísticas de rebatedores}
\keyword{datasets}
