\docType{data}
\name{salarios}
\alias{salarios}
\title{Tabela de salários}
\format{Um data.frame com 26.428 linhas e 5 colunas
\describe{
\item{ano}{Ano}
\item{id_equipe}{ID da equipe (fator)}
\item{id_liga}{ID da liga (fator)}
\item{id_jogador}{ID do jogador}
\item{salario}{Salário em dólares}
}}
\description{Dados salariais do jogador}
\keyword{datasets}
