\docType{data}
\name{tabela1}
\alias{tabela1}
\title{Registros de tuberculose da Organização Mundial da Saúde (primeira variante)}
\format{Um tibble com 6 linhas e 4 colunas
\describe{
\item{pais}{Nome do país}
\item{ano}{Ano correspondente aos valores registrados}
\item{casos}{Número de casos registrados}
\item{populacao}{População do país}
}}
\description{Contém observações de casos registrados no Afeganistão, Brasil e China nos anos de 1999 e 2000}
\keyword{datasets}
