\docType{data}
\name{tabela2}
\alias{tabela2}
\title{Registros de tuberculose da Organização Mundial da Saúde (segunda variante)}
\format{Um tibble com 12 linhas e 4 colunas
\describe{
\item{pais}{Nome do país}
\item{ano}{Ano correspondente aos valores registrados}
\item{tipo}{Categoria a que pertence a observação}
\item{contagem}{Quantidade de observações para cada tipo}
}}
\description{Contém observações de casos registrados no Afeganistão, Brasil e China nos anos de 1999 e 2000}
\keyword{datasets}
