\docType{data}
\name{tabela4a}
\alias{tabela4a}
\title{Registros de tuberculose da Organização Mundial da Saúde (variante 4a)}
\format{Um tibble com 3 linhas e 3 colunas
\describe{
\item{pais}{Nome do país}
\item{1999}{Casos registrados no ano 1999}
\item{2000}{Casos registrados no ano 2000}
}}
\description{Contém observações de casos registrados no Afeganistão, Brasil e China nos anos de 1999 e 2000}
\keyword{datasets}
