\docType{data}
\name{tabela5}
\alias{tabela5}
\title{Registros de tuberculose da Organização Mundial da Saúde (quinta variante)}
\format{Um tibble com 6 linhas e 4 colunas
\describe{
\item{pais}{Nome do país}
\item{seculo}{Século correspondente aos valores registrados}
\item{ano}{Ano correspondente aos valores registrados}
\item{taxa}{Taxa de casos de tuberculose com relação a população total}
}}
\description{Contém observações de casos registrados no Afeganistão, Brasil e China nos anos de 1999 e 2000}
\keyword{datasets}
