\docType{data}
\name{top100musicas}
\alias{top100musicas}
\title{Top 100 músicas da Billboard no ano 2000}
\format{Um tibble com 317 linhas e 79 colunas
\describe{
\item{artista}{Nome do(a) artista}
\item{musica}{Nome da música}
\item{data.adicionada}{Data em que a música entrou no Top 100}
\item{semana1}{Ranking da música em cada semana, após entrar no ranking}
\item{semana2}{Ranking da música em cada semana, após entrar no ranking}
\item{semana3}{Ranking da música em cada semana, após entrar no ranking}
\item{semana4}{Ranking da música em cada semana, após entrar no ranking}
\item{semana5}{Ranking da música em cada semana, após entrar no ranking}
\item{semana6}{Ranking da música em cada semana, após entrar no ranking}
\item{semana7}{Ranking da música em cada semana, após entrar no ranking}
\item{semana8}{Ranking da música em cada semana, após entrar no ranking}
\item{semana9}{Ranking da música em cada semana, após entrar no ranking}
\item{semana10}{Ranking da música em cada semana, após entrar no ranking}
\item{semana11}{Ranking da música em cada semana, após entrar no ranking}
\item{semana12}{Ranking da música em cada semana, após entrar no ranking}
\item{semana13}{Ranking da música em cada semana, após entrar no ranking}
\item{semana14}{Ranking da música em cada semana, após entrar no ranking}
\item{semana15}{Ranking da música em cada semana, após entrar no ranking}
\item{semana16}{Ranking da música em cada semana, após entrar no ranking}
\item{semana17}{Ranking da música em cada semana, após entrar no ranking}
\item{semana18}{Ranking da música em cada semana, após entrar no ranking}
\item{semana19}{Ranking da música em cada semana, após entrar no ranking}
\item{semana20}{Ranking da música em cada semana, após entrar no ranking}
\item{semana21}{Ranking da música em cada semana, após entrar no ranking}
\item{semana22}{Ranking da música em cada semana, após entrar no ranking}
\item{semana23}{Ranking da música em cada semana, após entrar no ranking}
\item{semana24}{Ranking da música em cada semana, após entrar no ranking}
\item{semana25}{Ranking da música em cada semana, após entrar no ranking}
\item{semana26}{Ranking da música em cada semana, após entrar no ranking}
\item{semana27}{Ranking da música em cada semana, após entrar no ranking}
\item{semana28}{Ranking da música em cada semana, após entrar no ranking}
\item{semana29}{Ranking da música em cada semana, após entrar no ranking}
\item{semana30}{Ranking da música em cada semana, após entrar no ranking}
\item{semana31}{Ranking da música em cada semana, após entrar no ranking}
\item{semana32}{Ranking da música em cada semana, após entrar no ranking}
\item{semana33}{Ranking da música em cada semana, após entrar no ranking}
\item{semana34}{Ranking da música em cada semana, após entrar no ranking}
\item{semana35}{Ranking da música em cada semana, após entrar no ranking}
\item{semana36}{Ranking da música em cada semana, após entrar no ranking}
\item{semana37}{Ranking da música em cada semana, após entrar no ranking}
\item{semana38}{Ranking da música em cada semana, após entrar no ranking}
\item{semana39}{Ranking da música em cada semana, após entrar no ranking}
\item{semana40}{Ranking da música em cada semana, após entrar no ranking}
\item{semana41}{Ranking da música em cada semana, após entrar no ranking}
\item{semana42}{Ranking da música em cada semana, após entrar no ranking}
\item{semana43}{Ranking da música em cada semana, após entrar no ranking}
\item{semana44}{Ranking da música em cada semana, após entrar no ranking}
\item{semana45}{Ranking da música em cada semana, após entrar no ranking}
\item{semana46}{Ranking da música em cada semana, após entrar no ranking}
\item{semana47}{Ranking da música em cada semana, após entrar no ranking}
\item{semana48}{Ranking da música em cada semana, após entrar no ranking}
\item{semana49}{Ranking da música em cada semana, após entrar no ranking}
\item{semana50}{Ranking da música em cada semana, após entrar no ranking}
\item{semana51}{Ranking da música em cada semana, após entrar no ranking}
\item{semana52}{Ranking da música em cada semana, após entrar no ranking}
\item{semana53}{Ranking da música em cada semana, após entrar no ranking}
\item{semana54}{Ranking da música em cada semana, após entrar no ranking}
\item{semana55}{Ranking da música em cada semana, após entrar no ranking}
\item{semana56}{Ranking da música em cada semana, após entrar no ranking}
\item{semana57}{Ranking da música em cada semana, após entrar no ranking}
\item{semana58}{Ranking da música em cada semana, após entrar no ranking}
\item{semana59}{Ranking da música em cada semana, após entrar no ranking}
\item{semana60}{Ranking da música em cada semana, após entrar no ranking}
\item{semana61}{Ranking da música em cada semana, após entrar no ranking}
\item{semana62}{Ranking da música em cada semana, após entrar no ranking}
\item{semana63}{Ranking da música em cada semana, após entrar no ranking}
\item{semana64}{Ranking da música em cada semana, após entrar no ranking}
\item{semana65}{Ranking da música em cada semana, após entrar no ranking}
\item{semana66}{Ranking da música em cada semana, após entrar no ranking}
\item{semana67}{Ranking da música em cada semana, após entrar no ranking}
\item{semana68}{Ranking da música em cada semana, após entrar no ranking}
\item{semana69}{Ranking da música em cada semana, após entrar no ranking}
\item{semana70}{Ranking da música em cada semana, após entrar no ranking}
\item{semana71}{Ranking da música em cada semana, após entrar no ranking}
\item{semana72}{Ranking da música em cada semana, após entrar no ranking}
\item{semana73}{Ranking da música em cada semana, após entrar no ranking}
\item{semana74}{Ranking da música em cada semana, após entrar no ranking}
\item{semana75}{Ranking da música em cada semana, após entrar no ranking}
\item{semana76}{Ranking da música em cada semana, após entrar no ranking}
}}
\description{Classificações de músicas no top 100 da Billboard no ano 2000}
\keyword{datasets}
