\docType{data}
\name{veiculos}
\alias{veiculos}
\title{Dados de economia de combustível}
\format{Um data.frame com 33442 linhas e 12 colunas
\describe{
\item{id}{Identificador único EPA}
\item{fabricante}{Fabricante}
\item{modelo}{Nome do modelo}
\item{ano}{Ano de fabricação}
\item{classe}{tipo de Automóveis}
\item{transmissao}{tipo de transmissão}
\item{tracao}{tipo de tração}
\item{cilindros}{número de cilindros}
\item{motor}{tamanho do motor do automóvel, em litros}
\item{combustível}{tipo de combustível}
\item{autopista}{milhas por galão de combustível na estrada}
\item{cidade}{milhas por galão de combustível na cidade}
}}
\usage{veiculos}
\description{Dados de economia de combustível da Agência de Proteção Ambiental(EPA) dos Estados Unidos, 1985-2015. Contém uma seleção de variáveis e não considera veículos com dados incompletos.}
\keyword{datasets}
