\docType{data}
\name{bebes}
\alias{bebes}
\title{Nomes de bebês}
\format{Um data.frame com 1.924.665 linhas e 5 colunas
\describe{
\item{ano}{Ano de registro}
\item{sexo}{Sexo do(a) bebê}
\item{nome}{Nome}
\item{FALSE}{Número de ocorrências}
\item{prop}{Proporção}
}}
\usage{bebes}
\description{Dados completos sobre nomes de bebês fornecidos pela Administração de Segurança Social dos EUA. Inclui todos os nomes com pelo menos 5 usos}
\seealso{\code{\link[babynames]{babynames}}}
\keyword{datasets}
