\docType{data}
\name{companhias_aereas}
\alias{companhias_aereas}
\title{Nomes de companhias aéreas}
\format{Um data.frame com 16 linhas e 2 colunas
\describe{
\item{companhia_aerea}{Abreviação de dois dígitos do nome da companhia aérea}
\item{nome}{Nome completo da companhia aérea.}
}}
\usage{companhias_aereas}
\description{Nomes das companhias aéreas e seus respectivos códigos de dois dígitos.}
\seealso{\code{\link[nycflights13]{airlines}}}
\keyword{datasets}
