% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{package}
\name{dados-package}
\alias{dados}
\alias{dados-package}
\title{dados: Translate Datasets to Portuguese}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rivaquiroga/dados}
  \item Report bugs at \url{https://github.com/rivaquiroga/dados/issues}
}

}
\author{
\strong{Maintainer}: Riva Quiroga \email{riva.quiroga@uc.cl} (\href{https://orcid.org/0000-0002-1147-4135}{ORCID})

}
\keyword{internal}
