% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{package}
\name{dados-package}
\alias{dados}
\alias{dados-package}
\title{dados: Translate Datasets to Portuguese}
\description{
Este pacote faz a tradução dos seguintes 
    conjuntos de dados: 'airlines', 'airports', 'AwardsManagers',
    'babynames', 'Batting', 'diamonds', 'faithful', 'fueleconomy',
    'Fielding', 'flights', 'gapminder', 'gss_cat', 'iris', 'Managers',
    'mpg', 'mtcars', 'atmos', 'penguins', 'People, 'Pitching', 'planes',
    'presidential', 'table1', 'table2', 'table3', 'table4a', 'table4b',
    'table5', 'vehicles', 'weather', 'who'.  English: It provides a
    Portuguese translated version of the datasets listed above.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cienciadedatos/dados}
  \item Report bugs at \url{https://github.com/cienciadedatos/dados/issues}
}

}
\author{
\strong{Maintainer}: Riva Quiroga \email{riva.quiroga@uc.cl} (\href{https://orcid.org/0000-0002-1147-4135}{ORCID})

Authors:
\itemize{
  \item Sara Mortara \email{saramortara@gmail.com} (\href{https://orcid.org/0000-0001-6221-7537}{ORCID})
  \item Beatriz Milz \email{beatriz.milz@usp.br} (\href{https://orcid.org/0000-0002-3064-4486}{ORCID})
  \item Andrea Sánchez-Tapia \email{andreasancheztapia@gmail.com} (\href{https://orcid.org/0000-0002-3521-4338}{ORCID})
  \item Alejandra Andrea Tapia Silva \email{alejandraandreatapiasilva@gmail.com} (\href{https://orcid.org/0000-0003-0762-7618}{ORCID})
  \item Beatriz Maurer Costa \email{beatrizmcosta@me.com}
  \item Jean Prado \email{jeangrprado@gmail.com} (\href{https://orcid.org/0000-0002-7928-774X}{ORCID})
  \item Renata Hirota \email{re.hirota@gmail.com} (\href{https://orcid.org/0000-0002-3589-3246}{ORCID})
}

}
\keyword{internal}
