\docType{data}
\name{dados_iris}
\alias{dados_iris}
\title{Dados sobre a flor de Íris por Edgar Anderson}
\format{Um data.frame com 150 linhas e 5 colunas
\describe{
\item{Comprimento.Sepala}{Comprimento da Sépala}
\item{Largura.Sepala}{Largura da Sépala}
\item{Comprimento.Petala}{Comprimento da Pétala}
\item{Largura.Petala}{Largura da Pétala}
\item{Especies}{A espécie correspondente da flor de Íris (setosa, versicolor, virginica)}
}}
\usage{dados_iris}
\description{O conjunto de dados sobre a flor de Íris fornece as medidas em centímetros das variáveis de comprimento e largura da sépala e comprimento e largura da pétala, respectivamente, para 50 flores de cada uma das 3 espécies de íris. As espécies são Iris setosa, versicolor e virginica.}
\seealso{\code{\link[datasets]{iris}}}
\keyword{datasets}
