\docType{data}
\name{dados_oms2}
\alias{dados_oms2}
\title{Dados de tuberculose da Organização Mundial da Saúde}
\format{Um tibble com 7240 linhas e 58 colunas
\describe{
\item{pais}{Nome do país}
\item{ano}{Ano}
\item{fpp_h_014}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), homens (h) de 0 a 14 anos (014)}
\item{fpp_h_1524}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), homens (h) de 15 a 24 anos (1524)}
\item{fpp_h_2534}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), homens (h) de 25 a 34 anos (2534)}
\item{fpp_h_3544}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), homens (h) de 35 a 44 anos (3544)}
\item{fpp_h_4554}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), homens (h) de 45 a 54 anos (4554)}
\item{fpp_h_5564}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), homens (h) de 55 a 64 anos (5564)}
\item{fpp_h_65}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), homens (h) de 65 ou mais anos}
\item{fpp_m_014}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), mulheres (m) de 0 a 14 anos (014)}
\item{fpp_m_1524}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), mulheres (m) de 15 a 24 anos (1524)}
\item{fpp_m_2534}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), mulheres (m) de 25 a 34 anos (2534)}
\item{fpp_m_3544}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), mulheres (m) de 35 a 44 anos (3544)}
\item{fpp_m_4554}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), mulheres (m) de 45 a 54 anos (4554)}
\item{fpp_m_5564}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), mulheres (m) de 55 a 64 anos (5564)}
\item{fpp_m_65}{Casos novos, diagnóstico frotis pulmonar positivo (fpp), mulheres (m) de 65 ou mais anos}
\item{fpn_h_014}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), homens (h) de 0 a 14 anos (014)}
\item{fpn_h_1524}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), homens (h) de 15 a 24 anos (1524)}
\item{fpn_h_2534}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), homens (h) de 25 a 34 anos (2534)}
\item{fpn_h_3544}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), homens (h) de 35 a 44 anos (3544)}
\item{fpn_h_4554}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), homens (h) de 45 a 54 anos (4554)}
\item{fpn_h_5564}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), homens (h) de 55 a 64 anos (5564)}
\item{fpn_h_65}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), homens (h) de 65 ou mais anos}
\item{fpn_m_014}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), mulheres (m) de 0 a 14 anos (014)}
\item{fpn_m_1524}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), mulheres (m) de 15 a 24 anos (1524)}
\item{fpn_m_2534}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), mulheres (m) de 25 a 34 anos (2534)}
\item{fpn_m_3544}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), mulheres (m) de 35 a 44 anos (3544)}
\item{fpn_m_4554}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), mulheres (m) de 45 a 54 anos (4554)}
\item{fpn_m_5564}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), mulheres (m) de 55 a 64 anos (5564)}
\item{fpn_m_65}{Casos novos, diagnóstico frotis pulmonar negativo (fpn), mulheres (m) de 65 ou mais anos}
\item{ep_h_014}{Casos novos, diagnóstico extrapulmonar (ep), homens (h) de 0 a 14 anos (014)}
\item{ep_h_1524}{Casos novos, diagnóstico extrapulmonar (ep), homens (h) de 15 a 24 anos (1524)}
\item{ep_h_2534}{Casos novos, diagnóstico extrapulmonar (ep), homens (h) de 25 a 34 anos (2534)}
\item{ep_h_3544}{Casos novos, diagnóstico extrapulmonar (ep), homens (h) de 35 a 44 anos (3544)}
\item{ep_h_4554}{Casos novos, diagnóstico extrapulmonar (ep), homens (h) de 45 a 54 anos (4554)}
\item{ep_h_5564}{Casos novos, diagnóstico extrapulmonar (ep), homens (h) de 55 a 64 anos (5564)}
\item{ep_h_65}{Casos novos, diagnóstico extrapulmonar (ep), homens (h) de 65 ou mais anos}
\item{ep_m_014}{Casos novos, diagnóstico extrapulmonar (ep), mulheres (m) de 0 a 14 anos (014)}
\item{ep_m_1524}{Casos novos, diagnóstico extrapulmonar (ep), mulheres (m) de 15 a 24 anos (1524)}
\item{ep_m_2534}{Casos novos, diagnóstico extrapulmonar (ep), mulheres (m) de 25 a 34 anos (2534)}
\item{ep_m_3544}{Casos novos, diagnóstico extrapulmonar (ep), mulheres (m) de 35 a 44 anos (3544)}
\item{ep_m_4554}{Casos novos, diagnóstico extrapulmonar (ep), mulheres (m) de 45 a 54 anos (4554)}
\item{ep_m_5564}{Casos novos, diagnóstico extrapulmonar (ep), mulheres (m) de 55 a 64 anos (5564)}
\item{ep_m_65}{Casos novos, diagnóstico extrapulmonar (ep), mulheres (m) de 65 ou mais anos}
\item{recaida_h_014}{Novas recaídas, homens (h) de 0 a 14 anos (014)}
\item{recaida_h_1524}{Novas recaídas, homens (h) de 15 a 24 anos (1524)}
\item{recaida_h_2534}{Novas recaídas, homens (h) de 25 a 34 anos (2534)}
\item{recaida_h_3544}{Novas recaídas, homens (h) de 35 a 44 anos (3544)}
\item{recaida_h_4554}{Novas recaídas, homens (h) de 45 a 54 anos (4554)}
\item{recaida_h_5564}{Novas recaídas, homens (h) de 55 a 64 anos (5564)}
\item{recaida_h_65}{Novas recaídas, homens (h) de 65 ou mais anos}
\item{recaida_m_014}{Novas recaídas, mulheres (m) de 0 a 14 anos (014)}
\item{recaida_m_1524}{Novas recaídas, mulheres (m) de 15 a 24 anos (1524)}
\item{recaida_m_2534}{Novas recaídas, mulheres (m) de 25 a 34 anos (2534)}
\item{recaida_m_3544}{Novas recaídas, mulheres (m) de 35 a 44 anos (3544)}
\item{recaida_m_4554}{Novas recaídas, mulheres (m) de 45 a 54 anos (4554)}
\item{recaida_m_5564}{Novas recaídas, mulheres (m) de 55 a 64 anos (5564)}
\item{recaida_m_65}{Novas recaídas, mulheres (m) de 65 ou mais anos}
}}
\usage{dados_oms2}
\description{Subconjunto de dados do relatório anual de tuberculose da Organização Mundial da Saúde. A tibble dados_oms2 é uma versão levemente modificada de dados_oms, ajustando as variaveis para serem mais consistentes, removendo iso2, iso3 e adicionando _ depois do sexo biológico}
\seealso{\code{\link[tidyr]{who2}}}
\keyword{datasets}
