\docType{data}
\name{dados_starwars}
\alias{dados_starwars}
\title{Personagens de Starwars}
\format{Um tibble com 87 linhas e 14 colunas
\describe{
\item{nome}{Nome}
\item{altura}{Altura em centímetros (cm)}
\item{massa}{Massa em quilogramas (kg)}
\item{cor_do_cabelo}{Cor do cabelo}
\item{cor_da_pele}{Cor da pele}
\item{cor_dos_olhos}{Cor dos olhos}
\item{ano_nascimento}{Ano de nascimento BBY (BBY = antes da batalha de Yavin)}
\item{sexo_biologico}{Sexo Biológico, sendo macho, fêmea, hermafrodita, ou nenhum (no caso dos Droides)}
\item{genero}{Gênero - O papel de gênero ou identidade de gênero da/do personagem conforme determinado por sua personalidade ou a maneira como foram programados (como no caso dos Droides)}
\item{planeta_natal}{Nome do planeta natal}
\item{especie}{Nome da espécie}
\item{filmes}{Lista de filmes em que o/a personagem apareceu}
\item{veiculos}{Lista de veículos que o/a personagem pilotou nos filmes}
\item{naves_espaciais}{Lista de naves espaciais que o/a personagem pilotou nos filmes}
}}
\description{Contém observações de personagens de Starwars. Os dados originais, da SWAPI - a API de Starwars,  https://swapi.dev/ - foram revisados para refletir pesquisas adicionais sobre gênero e sexo dos/das personagens.}
\seealso{\code{\link[dplyr]{starwars}}}
\keyword{datasets}
