\docType{data}
\name{diamante}
\alias{diamante}
\title{Preço de 50.000 diamantes}
\format{Um data.frame con 53.940 linhas y 10 colunas
\describe{
\item{preco}{Preço em dólares americanos ($326-$18,823)}
\item{quilate}{Peso do diamante (0.2-5.01)}
\item{corte}{Qualidade de corte (Justo, Bom, Muito Bom, Premium, Ideal)}
\item{cor}{Cor de diamante, de J (peor) a D (melhor)}
\item{claridade}{Quão claro é o diamante (I1 (peor), SI1, SI2, VS1, VS2, VVS1, VVS2, IF (melhor))}
\item{profundidade}{Porcentagem da profundidade total = z / mean(x, y) = 2 * z / (x + y) (43-79)}
\item{tabela}{Largura do topo do diamante em relação ao seu ponto mais largo (43-95)}
\item{x}{Largo em milímetros}
\item{y}{Largura em milímetros}
\item{z}{Profundidade em milímetros}
}}
\usage{diamante}
\description{Um conjunto de dados contendo os preços de quase 54.000 diamantes}
\keyword{datasets}
