\docType{data}
\name{jardineiros}
\alias{jardineiros}
\title{Tabela de estatísticas do jardineiro}
\format{Um data frame com 140.921 linhas e 18 colunas
\describe{
\item{id_jogador}{ID do jogador}
\item{id_ano}{Ano}
\item{ordem_equipe}{Ordem em que o jogador se moveu entre as equipes na mesma temporada}
\item{id_equipe}{ID da equipe (fator)}
\item{id_liga}{ID da liga (fator com níveis AA, AL, FL, NL, PL, UA)}
\item{posicao}{Posição do jogador defensivo}
\item{jogos}{Jogos}
\item{jogos_iniciados}{Jogos iniciados}
\item{inn_outs}{Tempo jogado pelo jogador expresso em outs}
\item{outs_feitos}{Outs feitas pelo jogador}
\item{assistencias}{Assistências realizadas pelo jogador}
\item{erros}{Erros cometidos pelo jogador}
\item{eliminacao_dupla}{Eliminação dupla em que o jogador esteve envolvido defensivamente}
\item{bolas_passadas}{Bolas passadas (aplicável a receptores)}
\item{lancamentos_desviados}{Wild pitches - arremessos ruins que permitem o avanço de corredores adversários (aplicável a arremessadores)}
\item{bases_roubadas}{Bases roubadas pelo oponente (aplicável aos receptor)}
\item{capturado_roubando}{Opositores apanhados em tentativa de roubo (aplicável a receptores)}
\item{taxa_zona}{Porcentagem de bolas capturadas por um defensor, em sua zona defensiva típica}
}}
\description{Estatísticas de defesa}
\seealso{\code{\link[Lahman]{Fielding}}}
\keyword{datasets}
