\docType{data}
\name{pessoas}
\alias{pessoas}
\title{Tabela de pessoas}
\format{Um data frame com 19.617 linhas e 26 colunas
\describe{
\item{id_jogador}{ID do jogador}
\item{ano_nascimento}{Ano de nascimento}
\item{mes_nascimento}{Mês de nascimento}
\item{dia_nascimento}{Dia de nascimento}
\item{pais_nascimento}{País de nascimento}
\item{estado_nascimento}{Estado de nascimento}
\item{cidade_nascimento}{Cidade de nascimento}
\item{ano_falecimento}{Ano de falecimento}
\item{mes_falecimento}{Mês de falecimento}
\item{dia_falecimento}{Dia de falecimento}
\item{pais_falecimento}{País de falecimento}
\item{estado_falecimento}{Estado de falecimento}
\item{cidade_falecimento}{Cidade de falecimento}
\item{nome}{Nome}
\item{sobrenome}{Sobrenome}
\item{nome_proprio}{Nome próprio}
\item{peso}{Peso em libras}
\item{altura}{Altura em polegadas}
\item{rebatedor}{Se, como rebatedor, você é destro (D) ou canhoto (C)}
\item{arremessador}{Se, como arremessador, você é destro (D) ou canhoto (C)}
\item{data_primeiro_jogo}{Data do primeiro jogo na primeira divisão}
\item{data_ultimo_jogo}{Data do último jogo na primeira divisão}
\item{id_retrosheet}{Código do jogador usado em Retrosheet (http://www.retrosheet.org/)}
\item{id_brw}{Código do jogador usado em Baseball Reference website (http://www.baseball-reference.com/)}
\item{data_falecimento}{Data de falecimento em formato yyyy-mm-dd}
\item{data_nascimento}{Data de nascimento em formato yyyy-mm-dd}
}}
\description{Nomes de jogadores, data de nascimento e informaçôes bibliográficas. Este arquivo é usado para obter detalhes sobre os jogadores listados em rebatedores, arremessadores e outros arquivos nos quais os jogadores são identificados por id_jogador}
\seealso{\code{\link[Lahman]{People}}}
\keyword{datasets}
