\docType{data}
\name{pixar_avalicao_publico}
\alias{pixar_avalicao_publico}
\title{Filmes da Pixar e avaliaçôes}
\format{Um tibble com 23 linhas e 5 colunas
\describe{
\item{filme}{Título do filme}
\item{nota_rotten_tomatoes}{Avaliação obtida pelo site Rotten Tomatoes. As avaliaçôes são numéricas e vão de 0 a 100.}
\item{nota_metacritic}{Avaliação obtida pelo site Metacritic. As avaliaçôes são numéricas e vão de 0 a 100.}
\item{nota_cinema_score}{Avaliação obtida pelo site CinemaScore. Avaliação em escala alfabética que varia de A+ a F.}
\item{nota_critics_choice}{Avaliação obtida pelo Critics' Choice Movie Awards que varia de 0 a 100.}
}}
\usage{pixar_avalicao_publico}
\description{Filmes da Pixar com avaliaçôes do público e da crítica, obtidas por diferentes sites}
\seealso{\code{\link[pixarfilms]{public_response}}}
\keyword{datasets}
