\docType{data}
\name{pixar_bilheteria}
\alias{pixar_bilheteria}
\title{Bilheteria dos filmes da Pixar}
\format{Um tibble com 23 linhas e 5 colunas
\describe{
\item{filme}{Título do filme}
\item{orcamento}{Orçamento do filme (em dólar americano)}
\item{bilheteria_eua_canada}{Valor gerado pela venda de ingressos nos EUA e Canada (em dólar americano)}
\item{bilheteria_outros_paises}{Valor gerado pela venda de ingressos nos demais países (em dólar americano)}
\item{bilheteria_mundial}{Valor gerado com a venda de ingressos no mundo inteiro (em dólar americano)}
}}
\usage{pixar_bilheteria}
\description{Filmes da Pixar com seus respectivos orçamentos e valor de bilheteria}
\seealso{\code{\link[pixarfilms]{box_office}}}
\keyword{datasets}
