\docType{data}
\name{pixar_equipe}
\alias{pixar_equipe}
\title{Equipe dos filmes}
\format{Um tibble com 228 linhas e 3 colunas
\describe{
\item{filme}{Título do filme}
\item{cargo}{Cargo das pessoas}
\item{nome}{Nome da pessoa que ocupa o cargo}
}}
\usage{pixar_equipe}
\description{Principais pessoas envolvidas nas produções dos filmes da Pixar}
\seealso{\code{\link[pixarfilms]{pixar_people}}}
\keyword{datasets}
