\docType{data}
\name{pixar_filmes}
\alias{pixar_filmes}
\title{Filmes da Pixar}
\format{Um tibble com 27 linhas e 3 colunas
\describe{
\item{ordem_lancamento}{Ordem de lançamento dos filmes}
\item{filme}{Título do filme}
\item{data_lancamento}{Data de lançamento do filme}
}}
\usage{pixar_filmes}
\description{Nome e data e ordem de lançamento dos filmes da Pixar}
\seealso{\code{\link[pixarfilms]{pixar_films}}}
\keyword{datasets}
