\docType{data}
\name{presidentes_eua}
\alias{presidentes_eua}
\title{Período que engloba 12 presidentes, desde Eisenhower até Trump}
\format{Um data.frame com 12 linhas e 4 colunas
\describe{
\item{nome}{Nome do presidente}
\item{inicio}{Data de início do período presidencial}
\item{fim}{Data do término do período presidencial}
\item{partido}{Partido político ao qual pertence}
}}
\description{Nome, data de início e de término do mandato e partido político de 12 presidentes dos Estados Unidos, de Eisenhower a Trump.}
\seealso{\code{\link[ggplot2]{presidential}}}
\keyword{datasets}
