\docType{data}
\name{questionario}
\alias{questionario}
\title{Amostra de variáveis categóricas do questinário 'General Social Survey' (GSS)}
\format{Um data.frame com 21.473 linhas e 9 colunas
\describe{
\item{ano}{Ano do questionário, 2000-2014}
\item{estado_civil}{Estado civil}
\item{idade}{Idade. Idade máxima limitada a 89 anos}
\item{raca}{Categorias raciais segundo o questionário}
\item{renda}{Renda respondida pelas pessoas entrevistadas (US$)}
\item{partido}{Partido político com quem a pessoa entrevistada se identifica}
\item{religiao}{Religião}
\item{denominacao}{Denominação religiosa a que pertence}
\item{horas_tv}{Horas de televisão assistidas por dia}
}}
\usage{questionario}
\description{Amostra de variáveis categóricas do questinário 'General Social Survey' (GSS)}
\seealso{\code{\link[forcats]{gss_cat}}}
\keyword{datasets}
