\docType{data}
\name{tabela1}
\alias{tabela1}
\title{Registros de tuberculose da Organização Mundial da Saúde (Primeira variante)}
\format{Um tibble com 6 linhas e 4 colunas
\describe{
\item{pais}{Nome do País}
\item{ano}{Ano correspondente aos valores registrados}
\item{casos}{Número de casos registrados}
\item{populacao}{População do país}
}}
\description{Contém observações de casos registrados no Afeganistão, Brasil e China nos anos 1999 e 2000.}
\seealso{\code{\link[tidyr]{table1}}}
\keyword{datasets}
