\docType{data}
\name{tabela3}
\alias{tabela3}
\title{Registros de tuberculose da Organização Mundial da Saúde (terceira variante)}
\format{Um tibble com 6 linhas e 3 colunas
\describe{
\item{pais}{Nome do país}
\item{ano}{Ano correspondente aos valores registrados}
\item{taxa}{Taxa de casos de tuberculose em relação a população total}
}}
\description{Contém observações de casos registrados no Afeganistão, Brasil e China nos anos de 1999 e 2000}
\seealso{\code{\link[tidyr]{table3}}}
\keyword{datasets}
