\docType{data}
\name{velho_fiel}
\alias{velho_fiel}
\title{Dados do gêiser Velho Fiel (Old Faithful)}
\format{Um data.frame com 272 linhas e 2 colunas
\describe{
\item{erupcoes}{Tempo das erupçôes (em minutos)}
\item{tempo_espera}{Minutos entre as erupçôes}
}}
\usage{velho_fiel}
\description{Dados do tempo de espera entre erupçôes e duração de cada erupção do gêiser Velho Fiel (Old Faithful), localizado no Parque Nacional de Yellowstone, em Wyoming, nos Estados Unidos.}
\seealso{\code{\link[datasets]{faithful}}}
\keyword{datasets}
